Fixed issue with read after write
per the open group documentation:

When a file is opened with update mode ('+' as the second or third
character in the mode argument), both input and output may be performed
on the associated stream. However, the application shall ensure that
output is not directly followed by input without an intervening call to
fflush() or to a file positioning function (fseek(), fsetpos(), or
rewind()).